requirements.txt
flask
chromadb
unstructured
langchain
langchain-text-splitters
unstructured[all-docs]
langchain-community
langchain-core
python-dotenv
sqlite3